implement authorization using Gate Facade

+++ app/Http/Controllers/QuestionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Question;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\AskQuestionRequest;
 use Illuminate\Contracts\Session\Session;
 
@@ -69,6 +70,9 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, "Access Denied");
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -81,6 +85,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, 'Access Denied');
+        }
+
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'Update question to database succesfully');
     }
@@ -93,6 +101,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403, 'Access Denied');
+        }
+
         $question->delete();
         
         return redirect()->route('questions.index')->with('success', 'Success delete question');
